factor out alg
